# docs/api/ui-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Issue Orchestrator UI API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/dialog/session-diagnostics/{issue_number}": {
      "get": {
        "summary": "Get session diagnostics dialog",
        "parameters": [
          {
            "name": "issue_number",
            "in": "path",
            "required": true,
            "schema": {"type": "integer"}
          }
        ],
        "responses": {
          "200": {
            "description": "Session diagnostics dialog",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/SessionDiagnosticsDialogPayload"}
              }
            }
          }
        }
      }
    },
    "/api/dialog/validation-failure/{issue_number}": {
      "get": {
        "summary": "Get validation failure dialog",
        "parameters": [
          {
            "name": "issue_number",
            "in": "path",
            "required": true,
            "schema": {"type": "integer"}
          }
        ],
        "responses": {
          "200": {
            "description": "Validation failure dialog",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ValidationFailureDialogPayload"}
              }
            }
          }
        }
      }
    },
...
    "/api/dialog/phase/{issue_number}": {
      "get": {
        "summary": "Get phase dialog",
        "parameters": [
          {
            "name": "issue_number",
            "in": "path",
            "required": true,
            "schema": {"type": "integer"}
          },
          {
            "name": "phase",
            "in": "query",
            "required": false,
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "Phase dialog",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/PhaseDialogPayload"}
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DialogRowPayload": {
        "type": "object",
        "additionalProperties": false,
        "required": ["label", "value"],
        "properties": {
          "label": {"type": "string"},
          "value": {"type": "string"}
        }
      },
...
      "SessionDiagnosticsActionPayload": {
        "type": "object",
        "additionalProperties": true,
        "required": ["type", "label"],
        "properties": {
          "type": {"type": "string"},
          "label": {"type": "string"},
          "path": {"type": ["string", "null"]},
          "issue_number": {"type": ["integer", "null"]},
          "group": {
            "type": ["string", "null"],
            "enum": ["validation_artifacts", "session_evidence", "diagnostics", null]
          }
        }
      },
      "SessionDiagnosticsDialogPayload": {
        "type": "object",
        "additionalProperties": false,
        "required": ["title", "rows", "actions"],
        "properties": {
          "title": {"type": "string"},
          "rows": {"type": "array", "items": {"$ref": "#/components/schemas/DialogRowPayload"}},
          "actions": {"type": "array", "items": {"$ref": "#/components/schemas/SessionDiagnosticsActionPayload"}},
          "analysis": {
            "oneOf": [
              {"$ref": "#/components/schemas/SessionDiagnosticsAnalysisPayload"},
              {"type": "null"}
            ]
          },
          "follow_up_issues": {
            "type": "array",
            "items": {"$ref": "#/components/schemas/SessionDiagnosticsFollowUpIssuePayload"}
          }
        }
      },
      "SessionDiagnosticsAnalysisPayload": {
        "type": "object",
        "additionalProperties": false,
        "required": ["headline"],
        "properties": {
          "headline": {"type": "string"},
          "detail": {"type": ["string", "null"]},
          "suggestions": {"type": "array", "items": {"type": "string"}}
        }
      },
      "SessionDiagnosticsFollowUpIssuePayload": {
        "type": "object",
        "additionalProperties": false,
        "required": ["title", "reason", "blocking"],
        "properties": {
          "title": {"type": "string"},
          "reason": {"type": "string"},
          "evidence": {"type": ["string", "null"]},
          "suggested_labels": {"type": "array", "items": {"type": "string"}},
          "blocking": {"type": "boolean"}
        }
      },
      "ValidationFailureActionSectionPayload": {
        "type": "object",
        "additionalProperties": false,
        "required": ["title", "actions"],
        "properties": {
          "title": {"type": "string"},
          "actions": {"type": "array", "items": {"$ref": "#/components/schemas/SessionDiagnosticsActionPayload"}}
        }
      },
      "JUnitCasePayload": {
        "type": "object",
        "additionalProperties": false,
        "required": ["case_id", "display_name", "outcome"],
        "properties": {
          "case_id": {"type": "string"},
          "display_name": {"type": "string"},
          "suite_name": {"type": ["string", "null"]},
          "outcome": {"type": "string", "enum": ["passed", "failed", "error", "skipped"]},
          "duration_seconds": {"type": ["number", "null"]},
          "failure_details": {"type": ["string", "null"]},
          "system_out": {"type": ["string", "null"]},
          "system_err": {"type": ["string", "null"]}
        }
      },
      "ValidationFailureDialogPayload": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "title",
          "status",
          "reason",
          "suite",
          "command",
          "exit_code",
          "started_at",
          "ended_at",
          "failed_tests",
          "stdout_excerpt",
          "stderr_excerpt",
          "junit_cases",
          "summary_rows",
          "action_sections"
        ],
        "properties": {
          "title": {"type": "string"},
          "status": {"type": "string", "enum": ["passed", "failed"]},
          "reason": {"type": "string"},
          "suite": {"type": "string"},
          "command": {"type": "string"},
          "exit_code": {"type": ["integer", "null"]},
          "started_at": {"type": "string"},
          "ended_at": {"type": "string"},
          "failed_tests": {"type": "array", "items": {"type": "string"}},
          "stdout_excerpt": {"type": "array", "items": {"type": "string"}},
          "stderr_excerpt": {"type": "array", "items": {"type": "string"}},
          "junit_cases": {"type": "array", "items": {"$ref": "#/components/schemas/JUnitCasePayload"}},
          "summary_rows": {"type": "array", "items": {"$ref": "#/components/schemas/DialogRowPayload"}},
          "action_sections": {"type": "array", "items": {"$ref": "#/components/schemas/ValidationFailureActionSectionPayload"}}
        }
      },
...
      "PhaseDialogPayload": {
        "type": "object",
        "additionalProperties": false,
        "required": ["title", "issue_number", "phase", "phases"],
        "properties": {
          "title": {"type": "string"},
          "issue_number": {"type": "integer"},
          "phase": {"type": ["object", "null"], "additionalProperties": true},
          "phases": {"type": "array", "items": {"type": "object", "additionalProperties": true}}
        }
      }
    }
  }
}